+++ packages/function-enhancers/src/index.ts
@@ -1,1 +1,3 @@
 export {default as memoize} from './memoize';
+export {default as debounce} from './debounce';
+export {default as throttle} from './throttle';